Easy Player Values + random enemy placement
I changed a bunch of code. I changed code around to make bullets controlled by the player script instead of the main game script. I also made it so you can control the speed of bullets using the BULLETSPEED value in the player.  I made a bullet2 object that switches the script to the root node because the script being on the Area2D was messing something up. Bullet  Cooldown doesn't work yet